SdmGatekeeper(): Still trying to debug the session loss problem, it seems to be related to the call to session_set_cookie_params() in the SdmGatekeeper::SessionStart() method, specifically the  parameter which should be set to our sites root url however setting  to the site root url seems to cause the session to become invalid. Working on a fix..., session_regenerate_id() has been determined NOT to be a culprit in our session data loss and is again apart of our SdmGatekeeper::sessionStart() method, SdmGatekeeper::sessionStart() now returns FALSE if either session_start() or session_regenerate_id() fails; reset.php: Added check to make sure reset.php does not try and delete any . or .. files when cleaning up the session_save_path() directory on site reset.

+++ reset.php
@@ -14,7 +14,9 @@ $sdmcore = new SdmCore();
 /** Delete any old session data */
 $targets = scandir(session_save_path());
 foreach ($targets as $sessfile) {
-    unlink(session_save_path() . '/' . $sessfile);
+    if ($sessfile != '.' && $sessfile != '..') {
+        unlink(session_save_path() . '/' . $sessfile);
+    }
 }
 /** Setup default menus for new site */
 /** Main Menu */
